-apply changes suggested by SonarCloud -remove Main class

+++ src/main/java/org/fundacionjala/coding/rmendoza/Amount.java
@@ -6,14 +6,14 @@ class Amount {
         // not called
     }
 
-    static final int NEW_RELEASE_DAILY_CHARGE = 3;
+    static final double NEW_RELEASE_DAILY_CHARGE = 3.0D;
     static final double REGULAR_DAILY_CHARGE = 1.5D;
     static final double CHILDREN_DAILY_CHARGE = 1.5D;
 
     static final int ALLOWED_REGULAR_RENTAL_DAYS = 2;
     static final int ALLOWED_CHILDREN_RENTAL_DAYS = 3;
 
-    static final int REGULAR_AMOUNT = 2;
+    static final double REGULAR_AMOUNT = 2.0D;
     static final double CHILDREN_AMOUNT = 1.5D;
 
     static final int NEW_RELEASE_RENTAL_DAY_BONUS = 2;
